requirements.txt
numpy
torch
opencv-python
pillow
moviepy
einops
transformers
safetensors
openai
aiohttp